+++ WebApp/main.js
@@ -1,0 +1,11 @@
+
+$(document).ready(function(){
+    // console.log("Hello World")
+    // token = Cookies.get("token") 
+    // if (token == undefined){
+    //     token = ""
+    // }
+    // file = "../fragen_abwl.xlsx"
+    // var workbook = XLSX.readFile(file);
+    
+});
