+++ WebApp/templates/main.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset='utf-8'>
+    <meta http-equiv='X-UA-Compatible' content='IE=edge'>
+    <title>ABWL & Recht Quiz</title>
+    <meta name='viewport' content='width=device-width, initial-scale=1'>
+    <link rel='stylesheet' type='text/css' media='screen' href='main.css'>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.1/jquery.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/js-cookie@3.0.1/dist/js.cookie.min.js"></script>
+    <script src="https://cdn.sheetjs.com/xlsx-latest/package/dist/xlsx.full.min.js"></script>
+    
+    <!-- <script src='main.js'></script> -->
+</head>
+<body>
+    <input id="username">
+    <button id="button">Start</button>
+
+    <script>
+        $('#button').click(function() {
+        var name = $('#username').val();
+        if (name != ""){
+            window.location.href = '/quiz/'+name;
+        } else {
+            alert("Username empty!")
+        }
+    }); 
+    </script>
+</html>
